requirements.txt
celery==4.1.0
coreapi==2.3.3
Django==1.10
django-filter==1.0.4
django-redis==4.8.0
djangorestframework==3.7.0
kombu==4.1.0
psycopg2==2.7.3.1
redis==2.10.6
requests==2.18.4
